+++ eras/conway/impl/CHANGELOG.md
@@ -50,6 +50,7 @@
 * Remove `DelegStakeTxCert` from the `COMPLETE` pragma for `TxCert`
 * Add `Committee` and adjust `NewCommittee` governance action
 * Add `ConwayUpdateDRep` constructor to `ConwayGovCert` type and corresponding pattern `UnRegDRepTxCert`
+* Update `ProposalProcedure` return address to be a `RewardAcnt`
 
 ## 1.6.3.0